+++ index.js
@@ -1,3 +1,3 @@
 'use strict';
 
-module.exports = require('./lib/egg_loader');
+module.exports = require('./lib/loader/egg_loader');
