+++ lib/loader/context_loader.js
@@ -2,9 +2,9 @@
 
 const assert = require('assert');
 const is = require('is-type-of');
-const Loader = require('./loader');
+const FileLoader = require('./file_loader');
 const classLoader = Symbol('classLoader');
-const EXPORTS = Loader.EXPORTS;
+const EXPORTS = FileLoader.EXPORTS;
 
 class ClassLoader {
 
@@ -31,7 +31,7 @@ class ClassLoader {
   }
 }
 
-class ContextLoader extends Loader {
+class ContextLoader extends FileLoader {
 
   constructor(options) {
     assert(options.field, 'options.field is required');
