+++ lib/loader/egg_loader.js
@@ -5,10 +5,10 @@ const path = require('path');
 const assert = require('assert');
 const isFunction = require('is-type-of').function;
 const debug = require('debug')('egg-loader');
-const Loader = require('./loader');
+const FileLoader = require('./file_loader');
 const ContextLoader = require('./context_loader');
-const loadFile = require('./utils').loadFile;
-const getHomedir = require('./utils').getHomedir;
+const loadFile = require('../utils').loadFile;
+const getHomedir = require('../utils').getHomedir;
 const Emitter = require('events').EventEmitter;
 
 class EggLoader {
@@ -243,7 +243,7 @@ class EggLoader {
       target,
       inject: this.app,
     }, opt);
-    new Loader(opt).load();
+    new FileLoader(opt).load();
   }
 
   loadToContext(directory, field, opt) {
