+++ lib/loader/file_loader.js
@@ -6,7 +6,7 @@ const debug = require('debug')('egg-loader:loader');
 const path = require('path');
 const globby = require('globby');
 const is = require('is-type-of');
-const loadFile = require('./utils').loadFile;
+const loadFile = require('../utils').loadFile;
 const FULLPATH = Symbol('EGG_LOADER_ITEM_FULLPATH');
 const EXPORTS = Symbol('EGG_LOADER_ITEM_EXPORTS');
 
@@ -21,7 +21,7 @@ const defaults = {
   inject: undefined,
 };
 
-class Loader {
+class FileLoader {
 
   constructor(options) {
     assert(options.directory, 'options.directory is required');
@@ -92,7 +92,7 @@ class Loader {
 
 }
 
-module.exports = Loader;
+module.exports = FileLoader;
 module.exports.EXPORTS = EXPORTS;
 
 // a/b/c.js => ['a', 'b', 'c']
