+++ lib/loader/mixin/config.js
