+++ lib/loader/mixin/controller.js
