+++ lib/loader/mixin/extend.js
@@ -4,7 +4,7 @@ const fs = require('fs');
 const path = require('path');
 const interopRequire = require('interop-require');
 const debug = require('debug')('egg-loader:extend');
-const utils = require('../utils');
+const utils = require('../../utils');
 
 module.exports = {
