+++ lib/loader/mixin/middleware.js
