+++ lib/loader/mixin/plugin.js
@@ -3,8 +3,8 @@
 const fs = require('fs');
 const path = require('path');
 const debug = require('debug')('egg-loader:plugin');
-const sequencify = require('../utils/sequencify');
-const loadFile = require('../utils').loadFile;
+const sequencify = require('../../utils/sequencify');
+const loadFile = require('../../utils').loadFile;
 
 module.exports = {
