+++ lib/loader/mixin/proxy.js
