+++ lib/loader/mixin/service.js
